+++ src/main/java/bit/user/domain/User.java
@@ -17,7 +17,7 @@ import java.util.List;
 @Getter
 @Builder
 @AllArgsConstructor
-public class User implements UserDetails {
+public class User {
     private final Long id;
     private final String email;
     private final String nickName;
@@ -46,39 +46,4 @@ public class User implements UserDetails {
                 .build();
     }
 
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return UserDetails.super.isAccountNonExpired();
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return UserDetails.super.isAccountNonLocked();
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return UserDetails.super.isCredentialsNonExpired();
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return UserDetails.super.isEnabled();
-    }
-
-    @Override // 권한 반환
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return List.of(new SimpleGrantedAuthority("user"));
-    }
-
-    @Override
-    public String getPassword() {
-        return "";
-    }
-
-    @Override
-    public String getUsername() {
-        return email;
-    }
 }
