+++ src/main/java/bit/user/service/UserDetailService.java
@@ -1,5 +1,7 @@
 package bit.user.service;
 
+import bit.auth.domain.UserPrincipal;
+import bit.user.domain.User;
 import bit.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -17,6 +19,8 @@ public class UserDetailService implements UserDetailsService {
     // 사용자 이름(email)으로 사용자의 정보를 가져오는 메소드
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        return userRepository.findByEmail(email).orElseThrow(() -> new IllegalArgumentException(email));
+        User user = userRepository.findByEmail(email).orElseThrow(() -> new IllegalArgumentException(email));
+
+        return new UserPrincipal(user);
     }
 }
